# static/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoOnCall",
    "description": "基于 LangChain 的智能oncall运维系统",
    "version": "1.2.1"
  },
  "paths": {
...
    "/api/aiops/tools/contracts": {
      "get": {
        "tags": [
          "AIOps智能运维"
        ],
        "summary": "List Aiops Tool Contracts",
        "description": "Return read-only AIOps tool contracts without invoking external systems.",
        "operationId": "list_aiops_tool_contracts_api_aiops_tools_contracts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Aiops Tool Contracts Api Aiops Tools Contracts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/aiops/demo/incidents": {
      "get": {
        "tags": [
          "AIOps智能运维"
        ],
        "summary": "List Demo Incidents",
        "description": "Return the central demo incident catalog used by the frontend workbench.",
        "operationId": "list_demo_incidents_api_aiops_demo_incidents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Demo Incidents Api Aiops Demo Incidents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/aiops/status-catalog": {
      "get": {
        "tags": [
          "AIOps智能运维"
        ],
        "summary": "Get Aiops Status Catalog",
        "description": "Return lifecycle statuses used by AIOps history filters and badges.",
        "operationId": "get_aiops_status_catalog_api_aiops_status_catalog_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Aiops Status Catalog Api Aiops Status Catalog Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/aiops/runs": {
      "get": {
        "tags": [
          "AIOps智能运维"
        ],
        "summary": "List Aiops Runs",
        "description": "Return recent durable diagnosis runs for history views.",
        "operationId": "list_aiops_runs_api_aiops_runs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Incident Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "service_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Aiops Runs Api Aiops Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/aiops/runs/{session_id}": {
      "get": {
        "tags": [
          "AIOps智能运维"
        ],
        "summary": "Get Aiops Run Status",
        "description": "Return the latest durable state for one diagnosis run.",
        "operationId": "get_aiops_run_status_api_aiops_runs__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Session Id"
            }
          },
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Aiops Run Status Api Aiops Runs  Session Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/aiops": {
      "post": {
        "tags": [
          "AIOps智能运维"
        ],
        "summary": "Diagnose Stream",
        "description": "Run the Plan-Execute-Replan AIOps diagnosis workflow as an SSE stream.",
        "operationId": "diagnose_stream_api_aiops_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIOpsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/aiops/demo/incidents/{case_id}": {
      "get": {
        "tags": [
          "AIOps智能运维"
        ],
        "summary": "Get Demo Incident",
        "description": "Return a ready-to-run demo incident payload for interviews and local demos.",
        "operationId": "get_demo_incident_api_aiops_demo_incidents__case_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Case Id"
            }
          },
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/aiops/demo/incidents/{case_id}/run": {
      "post": {
        "tags": [
          "AIOps智能运维"
        ],
        "summary": "Run Demo Incident",
        "description": "Run a fixed demo incident through the normal AIOps SSE workflow.",
        "operationId": "run_demo_incident_api_aiops_demo_incidents__case_id__run_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Case Id"
            }
          },
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AIOpsRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{incident_id}/diagnosis/resume": {
      "post": {
        "tags": [
          "AIOps智能运维"
        ],
        "summary": "Resume Diagnosis Stream",
        "description": "Record an approved human decision and close the paused diagnosis loop.",
        "operationId": "resume_diagnosis_stream_api_incidents__incident_id__diagnosis_resume_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Incident Id"
            }
          },
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIOpsResumeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{incident_id}/changes/{change_plan_id}/resume": {
      "post": {
        "tags": [
          "AIOps智能运维"
        ],
        "summary": "Resume Safe Change Stream",
        "description": "Start the safe change workflow after an approval decision.",
        "operationId": "resume_safe_change_stream_api_incidents__incident_id__changes__change_plan_id__resume_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Incident Id"
            }
          },
          {
            "name": "change_plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Change Plan Id"
            }
          },
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeResumeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIOpsRequest": {
        "properties": {
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "会话ID，用于追踪诊断历史"
          },
          "incident": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Incident"
              },
              {
                "type": "null"
              }
            ],
            "description": "结构化故障事件；不传时后端会根据当前任务自动构造默认 Incident"
          }
        },
        "type": "object",
        "title": "AIOpsRequest",
        "description": "Request body for the streaming `/api/aiops` diagnosis endpoint.",
        "example": {
          "incident": {
            "environment": "prod",
            "service_name": "order-service",
            "severity": "P1",
            "symptom": "5xx 错误率升高，P95 延迟超过 3 秒，并出现 Redis connection timeout",
            "title": "order-service Redis timeout"
          },
          "session_id": "session-123"
        }
      },
      "AIOpsResumeRequest": {
        "properties": {
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "原诊断会话 ID；新审批会从 metadata 自动推断，旧审批可显式传入。"
          },
          "approval_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Approval Id",
            "description": "要恢复的审批 ID。"
          }
        },
        "type": "object",
        "required": [
          "approval_id"
        ],
        "title": "AIOpsResumeRequest",
        "description": "Request body for resuming a paused AIOps diagnosis after approval.",
        "example": {
          "approval_id": "apr_xxx",
          "session_id": "session-123"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}